// popover/types.ts
export type Triggers = "hover" | "click";

export type Behavior = "dropdown" | "multiple" | "single";

export enum Placement {
  default = "default",
  absolute = "absolute",
  fixed = "fixed",
}

export enum Position {
  left = "left",
  right = "right",
  top = "top",
  bottom = "bottom",
  rightTop = "rightTop",
  leftTop = "leftTop",
  rightBottom = "rightBottom",
  leftBottom = "leftBottom",
  inline = "inline",
  center = "center",
}

export enum Feature {
  smarty = "smarty",
}